### 2020-11-05 #2819 @whedon
[PRE REVIEW]: AltaiPony - Flare science in Kepler, K2 and TESS light curves
**Submitting author:** @ekaterinailin (<a href="http://orcid.org/0000-0002-6299-7542">Ekaterina Ilin</a>)
**Repository:** <a href="https://github.com/ekaterinailin/AltaiPony" target ="_blank">https://github.com/ekaterinailin/AltaiPony</a>
**Version:** v1.0
