### 2020-11-05 #2819 @ekaterinailin
> @ekaterinailin thanks for this submission. I'll be helping to find a handling editor.
>
> In the mean time, can you clarify if this submission is associated with an AAS paper?

No, it is not associated with an AAS paper.

Working up from the bottom of the reviewers list for suggestions: dflemin3, christinahedges
