### 2020-11-18 #2845 @mrtommyb
Hi @ekaterinailin I'm starting my review and working my way down the checklist. I'm at the item on Contribution and authorship. I have a question here.
I notice that the software Appaloosa is described as this software’s predecessor and it looks like some of the code in this package is lifted from Appaloosa. I was wondering if you could explain a little about your authorship choices for this paper given the heritage of some of the code included.
