### 2020-11-18 #2845 @ekaterinailin
Hi @mrtommyb ! First of all, thank you for volunteering as a reviewer.

My reasoning on authorship was: I made what I thought was a default choice based on the Github history of the project. From Appaloosa, I ended up using and adapting three functions (the aflare model, and the flare search functions in the residual light curve). AltaiPony was written from scratch, and did not grow out of Appaloosa directly, so maybe _predecessor_ is not the right term to use in a technical sense, uncertain here. In the end, I decided to cite Jim's code in the paper, as I did with K2SC and lightkurve, from which I integrated functions and classes in a similar manner.

Is there a consideration that I am missing here?
