### 2020-12-01 #2845 @afeinstein20
Hi @ekaterinailin ! Thanks for writing AltaiPony and what a fantastic logo! The software was easy to install and run, with few to minor errors. I've recorded some of those below as well as some changes I think would make the functionality and the documentation better/easier for users!

**Installation**:
I had no trouble installing and using the software following both installation methods. However, I did receive the following errors that I wanted to note. And there are some dependencies I found were missing later one.
1. Through PyPI, I receive the following error: `FileNotFoundError: [Errno 2] No such file or directory: '//anaconda3/lib/python3.7/site-packages/K2SC-1.0-py3.7.egg'`
2. Through Github clone, I receive the following error: `error: scipy 1.4.1 is installed but scipy!=1.4.0,!=1.4.1,>=0.19.0 is required by {'lightkurve'}`
3. I ran into trouble with the progressbar on my personal machine! I assume this is a version error because it runs okay in a Google Colab notebook. Error: `TypeError: __init__() got an unexpected keyword argument 'max_value'`
4. I went to run the FFD functionality and was missing the `emcee` and `corner` dependencies.

**Functionality**
1. It's unclear to me what the flare detection criteria are and if there's any way for me to change them. Are there any criteria I really shouldn't change? It looks like those criteria are stored in `find_flares_in_cont_obs_period` but there's no documentation for this function.
2. The injection-recovery tools and associated plots are super useful! It would be great to have a way to save these output figures, maybe through a keyword or by returning the figure instead of just plotting.
3. When I ran `simple_ffd.ed_and_freq()` from the documentation, it returned only 5 data points? I copied most of the lines from the demo notebooks in your documentation, so this is kind of weird!

**Documentation**
1. Could use a bit more description in the opening about paragraph, since there's way more to this software than just "finding flares" (e.g. light curve downloading, detrending, flare fitting, injection-recovery tests, etc.)
2. Example usage could use some figures! As I was walking through the tutorials, it would have been nice to compare what my light curve looked like compared to what the docs expected.

**Software Paper**
The paper is overall clear and well written. It provides a good outline as to why the software is necessary and what it accomplishes compared to other flare detection methods. I believe the paper would benefit from a bit more description about details of flare detection and fitting: What is detected as a flare and not a flare? What are the default settings for flare detection? How the software handle the ramp ups at the end of TESS orbits? What is defined as a the start and stop time of a flare? Adding little details like this would be useful!
