### 2020-12-01 #2845 @ekaterinailin
Hi @afeinstein20 ! Thanks a ton for the detailed review, these are all very helpful comments, and I am looking forward to addressing them to improve the software. I have a few busy weeks ahead of me, but I hope to get my hands on the code before the end of this year.
