### 2020-12-29 6e0647e
deleted JOSS paper draft and bibfile from this branch.
