+++ paper.bib
@@ -1,39 +1,0 @@
-@ARTICLE{ilin2020,
-       author = {{Ilin}, Ekaterina and {Schmidt}, Sarah J. and {Poppenh{\"a}ger}, Katja and
-         {Davenport}, James R.~A. and {Kristiansen}, Martti H. and
-         {Omohundro}, Mark},
-        title = "{Flares in Open Clusters with K2. II. Pleiades, Hyades, Praesepe, Ruprecht 147, and M67}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Solar and Stellar Astrophysics},
-         year = 2020,
-        month = oct,
-          eid = {arXiv:2010.05576},
-        pages = {arXiv:2010.05576},
-archivePrefix = {arXiv},
-       eprint = {2010.05576},
- primaryClass = {astro-ph.SR},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv201005576I},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-
-
-
-@ARTICLE{ramsay2020,
-       author = {{Ramsay}, Gavin and {Doyle}, J. Gerry and {Doyle}, Lauren},
-        title = "{TESS observations of southern ultrafast rotating low-mass stars}",
-      journal = {\mnras},
-     keywords = {stars: activity, stars: flare, stars: late-type, stars: low-mass, stars: rotation, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Earth and Planetary Astrophysics},
-         year = 2020,
-        month = jul,
-       volume = {497},
-       number = {2},
-        pages = {2320-2326},
-          doi = {10.1093/mnras/staa2021},
-archivePrefix = {arXiv},
-       eprint = {2007.04071},
- primaryClass = {astro-ph.SR},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.497.2320R},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
