### 2020-12-29 08b19a6
Adding more options for flare candidate definition and documenting everything in the tutorial.
