### 2020-12-29 0f3e7b6
Improving find flares tutorial
