### 2020-12-29 070c3f2
plotting heatmaps now returns a figure. updated notebooks
