### 2020-12-29 2ecfef3
more guidance on all the stuff you can do with AltaiPony
