### 2020-12-29 3500aca
more guidance on all the stuff you can do with AltaiPony
