### 2020-12-29 95cc7a3
added some figures to finding data tutorial
