### 2020-12-29 d5e1091
added figures to finding flares
