### 2020-12-29 2d77180
added a table screen shot to flare characterization, fixed some typos in the docs
