+++ docs/source/install.rst
@@ -26,7 +26,7 @@ This package depends, on `lightkurve`, `k2sc`, `numpy`, `pandas` and some other 
 Getting Started
 ^^^^^^^^^^^^^^^^
 
-Working with Kepler and TESS light curves is similar, K2 light curves need extra attention, so we treat them separately. For each mission, there is a notebook that will guide you through the basic applications. Taking a look at the `Finding Data`_
+Working with Kepler and TESS light curves is similar, K2 light curves need extra attention, so we treat them separately. For each mission, there is a notebook that will guide you through the basic applications. We recommend taking a look at the `Finding Data`_ tutorial.
 
 Kepler Light Curves
 ...................
@@ -56,12 +56,12 @@ Once you have tried basic **AltaiPony** on your light curves, you can start to a
 Test the performance of your flare finding algorithm
 .....................................................
 
-You may then want to test the perfomance of your chosen flare finding setup by injecting and recoving synthetic flares into your light curves. **AltaiPony** provides a framework to do so, explained in the `Synthetic Flare Injection and Recovery`_ tutorial. Check out the `Visualize_Injection_Recovery`_ notebook for nice plots.
+You may then want to test the perfomance of your chosen flare finding setup by injecting and recoving synthetic flares into your light curves. **AltaiPony** provides a framework to do so, explained in the `Synthetic Flare Injection and Recovery`_ tutorial. Check out the `visualization`_ notebook for nice plots.
 
 Analyze flare frequency distributions
 ......................................
 
-For a statistical analysis of your flares, **AltaiPony** also features a set of tools for the analysis of flare frequency distributions, including visualization, and different methods for power law fitting. For starters, check out the tutorial on `Flare Frequency Distributions and Power Laws`_. If you want to go hands on, start with the `beginner`_ notebook. For more advanced application, like working with samples of multiple stars and their flares, go to `advanced`_. 
+For a statistical analysis of your flares, **AltaiPony** also features a set of tools for the analysis of flare frequency distributions, including visualization, and different methods for power law fitting. For starters, check out the tutorial on `Flare Frequency Distributions and Power Laws`_. If you want to go hands on, start with the `beginner`_ notebook. For more advanced applications, like working with samples of multiple stars and their flares, go to the `advanced`_ notebook. 
 
 
 .. _Aigrain et al. 2016: http://ascl.net/1605.012
@@ -72,9 +72,10 @@ For a statistical analysis of your flares, **AltaiPony** also features a set of 
 .. _scipy: https://docs.scipy.org/doc/scipy-0.15.1/reference/generated/scipy.signal.savgol_filter.html
 .. _other: https://github.com/ekaterinailin/AltaiPony/blob/master/notebooks/TESS_Light_Curves_With_Flares.ipynb
 .. _in this tutorial: https://altaipony.readthedocs.io/en/latest/tutorials/altai.html
+.. _Finding Flares: https://altaipony.readthedocs.io/en/latest/tutorials/altai.html
 .. _Finding Data: https://altaipony.readthedocs.io/en/latest/tutorials/lcio.html
 .. _Synthetic Flare Injection and Recovery: https://altaipony.readthedocs.io/en/latest/tutorials/fakeflares.html
-.. _Visualize_Injection_Recovery.ipynb: https://github.com/ekaterinailin/AltaiPony/blob/master/notebooks/Visualize_Injection_Recovery.ipynb
+.. _visualization: https://github.com/ekaterinailin/AltaiPony/blob/master/notebooks/Visualize_Injection_Recovery.ipynb
[…]
