### 2020-12-30 2433f01
Added a few words
... to make definition of flare candidate more clear.
