### 2020-12-30 718c1dc
bumped versions of lightkurve and scipy
