+++ setup.py
@@ -14,7 +14,7 @@ setuptools.setup(
     url="https://github.com/ekaterinailin/AltaiPony",
     packages=setuptools.find_packages(),
     install_requires = ['numpy>=1.15.1', 'pybind11','lightkurve==1.9.1','pandas>=1.0',
-                        'progressbar2','seaborn',  'k2sc==1.0.1.4'],
+                        'progressbar2','seaborn',  'k2sc==1.0.1.4', 'emcee','corner'],
     include_package_data=True,
     package_data={
       'altaipony': ['static/*csv']},
