+++ requirements.txt
@@ -1,8 +1,8 @@
 numpy>=1.15.1
 pandas>=1.0.
-scipy==1.3.0
+scipy>=1.3.0
 pybind11
-lightkurve==1.9.1
+lightkurve>=1.9.1
 k2sc==1.0.1.4
 emcee
 corner
