### 2020-12-30 2be20c3
pandas 1.1.5. crashes heatmap from pivot_table
