+++ requirements.txt
@@ -1,5 +1,5 @@
 numpy>=1.15.1
-pandas>=1.0.
+pandas>=1.0, !=1.1.5
 scipy>=1.3.0
 pybind11
 lightkurve>=1.9.1
