### 2020-12-30 074ca43
updated requirements, added FAQs
