### 2020-12-30 #2845 @ekaterinailin
Hello again,

I finally got my hands on the code again, so here are my (proto-)solutions to @afeinstein20 's review:

- **Installation**
    - I had no [...] later on.
        - Through PyPI, I receive the following error: FileNotFoundError: [Errno 2] No such file or directory: '//anaconda3/lib/python3.7/site-packages/K2SC-1.0-py3.7.egg'
            - Not sure about that one because I am not familiar with the anaconda + python3.7 combo. The released version runs on python3.6. It would be awesome if you could open an issue on that for the next release (which will  probably be on python3.7, and K2SC is still going through some updates.
        - Through Github clone, I receive the following error: error: scipy 1.4.1 is installed but scipy!=1.4.0,!=1.4.1,>=0.19.0 is required by {'lightkurve'}
            - Hm. I updated a few dependencies now, maybe this solves the problem. At least there were no restrictions on scipy being the 1.4.1. version in requirements.txt or setup.py. I added the requirement to use scipy>1.5 in hope of improvement.
        - I ran into trouble with the progressbar on my personal machine! I
assume this is a version error because it runs okay in a Google Colab
notebook. Error: TypeError: __init__() got an unexpected keyword argument 'max_value'
            - Had this problem before, but was not sure where it came from.. It's perhaps to do with the default version fetched, I hope my attempts to solve the problem in [this issue](https://github.com/ekaterinailin/AltaiPony/issues/55) will help.
        - I went to run the FFD functionality and was missing the emcee and corner dependencies.
            - Ah, thanks! Added them to setup.py.
- **Functionality**
    - It's unclear to me what the flare detection criteria are and if there's any way for me to change them. Are there any criteria I really shouldn't change? It looks like those criteria are stored in find_flares_in_cont_obs_period but there's no documentation for this function.
        - I added some more docs [here](https://altaipony.readthedocs.io/en/latest/tutorials/altai.html) to explain everything. Short answer: you can change everything now, but the default is a good heuristic in general, a bit generous perhaps.
    - The injection-recovery tools and associated plots are super useful! It would be great to have a way to save these output figures, maybe through a keyword or by returning the figure instead of just plotting.
        - Good idea! `FlareLightCurve.plot_ed_ratio_heatmap()` and `FlareLightCurve.plot_recovery_probability_heatmap()`  now both return a matplotlib figure. Updated the `Visualize_Injection_Recovery.ipynb` accordingly.
    - When I ran simple_ffd.ed_and_freq() from the documentation, it returned only 5 data points? I copied most of the lines from the demo notebooks in your documentation, so this is kind of  weird!
        - You mean [here](https://altaipony.readthedocs.io/en/latest/tutorials/ffds.html)? If your flc.flares table is short, this can totally be the case. But if you used the setup in ``Beginner_Flare_Frequency_Distributions_and_Power_Laws.ipynb`` this should not happen, because the toy example generates 200 events. I re-ran the notebook, and it looked fine. It would be cool, if you could open an issue with the script you used that failed to reproduce the example.
- **Documentation**
    - Could use a bit more description in the opening about paragraph, since there's way more to this software than just "finding flares" (e.g. light curve downloading, detrending, flare fitting, injection-recovery tests, etc.)
        - New description in the about section, and more details in [Getting Started](https://altaipony.readthedocs.io/en/latest/install.html#getting-started) added!
    - Example usage could use some figures! As I was walking through the tutorials, it would have been nice to compare what my light curve looked like compared to what the docs expected.
        - Added light curves and an illustrative table to Finding Data, [Finding Flares](https://altaipony.readthedocs.io/en/latest/tutorials/altai.html)
        - Added heatmaps and an illustrative table to [Injection Recovery](https://altaipony.readthedocs.io/en/latest/tutorials/lcio.html).
- **Software Paper**
    - The paper [...] would benefit from a bit more description about details of flare detection and fitting:
        - I think these are questions that users will regularly have, and since I expect them to go to the documentation for this, I opened a [FAQ section](https://altaipony.readthedocs.io/en/latest/faq.html) in the docs that addresses them in some detail. I only added a few words to the paper to make the definition of flare candidates more clear.
        - What is detected as a flare and not a flare?
            - I added a section to the [Finding Flares](https://altaipony.readthedocs.io/en/latest/tutorials/altai.html#defining-flare-candidates) tutorial that goes into detail here, and I link to it in the FAQs.
        - What are the default settings for flare detection?
            - Also answered in [Finding Flares](https://altaipony.readthedocs.io/en/latest/tutorials/altai.html#defining-flare-candidates), answers linked in FAQs.
        - How the software handle the ramp ups at the end of TESS orbits?
            - It doesn't unless your de-trending algorithm is tailored to do it. I added instructions on how to pass a custom de-trending function to `FlareLightCurve.find_flares()
        - What is defined as a the start and stop time of a flare?
            - Answered this in the FAQs, and added an example on how to mask flares based on flare start and stop indices.

Let me know if I understood your remarks correctly! Happy new year :smiley_cat:
