### 2021-01-26 3e1e47d
New feature: add decay phase with reduced threshold.
