+++ altaipony/tests/test_utils.py
@@ -1,8 +1,32 @@
+
 import pytest
 
 import numpy as np
+
 from ..flarelc import FlareLightCurve
-from ..utils import k2sc_quality_cuts, split_gaps
+from ..utils import k2sc_quality_cuts, split_gaps, expand_mask
+
+import copy
+
+def test_expand_mask():
+    # set up a test array with a single outlier and a sequence
+    a = np.ones(25).astype(int)
+    a[3] = 0
+    a[[15,16]] = 0
+
+    # check result
+    assert (expand_mask(copy.deepcopy(a)) == np.array([1, 1, 1, 0, 1,
+                                        1, 1, 1, 1, 1,
+                                        1, 1, 0, 0, 0,
+                                        0, 0, 0, 0, 0,
+                                        0, 0, 0, 1, 1])).all()
+
+    # this breaks the function:
+    a[7] = np.nan
+
+    # check that it does
+    with pytest.raises(ValueError) as e:
+        expand_mask(copy.deepcopy(a))
 
 
 def test_k2sc_quality_cuts():
