### 2021-02-04 #2845 @afeinstein20
Hi @ekaterinailin ! Thanks for addressing all of the points! I'm very happy with the updates 😄  The additional docs and FAQ section are super useful. I just have 2 additional comments/issues:

 * Good idea! `FlareLightCurve.plot_ed_ratio_heatmap()` and `FlareLightCurve.plot_recovery_probability_heatmap()`  now both return a matplotlib figure. Updated the `Visualize_Injection_Recovery.ipynb` accordingly.
- I'm glad you agreed this would be useful! Unfortunately, now I'm running into some issues the way the values are being binned. I've been running everything in this [Google Colab](https://colab.research.google.com/drive/1xmPHhHI4mpGcq01wlS1n6u07gtmpDt4V?usp=sharing) notebook. I think if you open it, the error messages should appear as well?

_* You mean [here](https://altaipony.readthedocs.io/en/latest/tutorials/ffds.html)? ... example._
- Yup this was totally the issue! I thought in your docs you were computing the FFD for TIC 29780677. But I know realize it was for GJ 1243, although I'm still getting fewer points still. This is probably because I've set `iterations=50` for the `flc.sample_flare_recovery()` just for time's sake. Does that sound right to you?

I've gone ahead and checked all of the boxes on my review list, since you've covered everything on my end! Hopefully these 2 minor things listed above are easily fixed!
