### 2021-02-08 #2845 @mrtommyb
Sorry that it took me a while to get back to this. I reinstalled the package, and I installed from the github repo. A few things to note
* it would be useful if altaipony.__version__ returned a version number.
* I followed the getting started notebook and the from_mast function doesn't work for me.
```python
TypeError: __init__() got an unexpected keyword argument '_required_columns'
```
I think the issue is that altaipony is incompatible with lightkurve version 2.x. I think that this should probably be fixed before this is accepted.

* I was trying to work out if I could download the data and then import that data into a FlareLightCurve object but I could not work out how to do that. This meant that I could not test the additional functionality.
* I didn't see any information on how to contribute to the project.
