### 2021-02-25 1552d54
more kwargs functionality, better heatmap labeling, shorter outlier padding
