+++ requirements.txt
@@ -3,6 +3,7 @@ pandas>=1.0, !=1.1.5
 scipy>=1.5.0
 pybind11
 lightkurve>=1.9.1
+astropy>=4.1
 k2sc==1.0.1.4
 emcee
 corner
