### 2021-02-25 f5b252f
new lightkurve dies not work with k2sc file, so fix an older version
