### 2021-02-25 49426ac
Addtails: Better capture of flare decay, small upgrades in layout and kwargs, fixed lightkurve version (#65)
* New feature: add decay phase with reduced threshold.

* Added a section in the docs for the tails

* added some links

* tested mask expansion

* added the actual function

* more kwargs functionality, better heatmap labeling, shorter outlier padding

* bump astropy

* new lightkurve dies not work with k2sc file, so fix an older version

* fixed tests
