### 2021-02-25 #2845 @ekaterinailin
You're right, @mrtommyb , the new version of lightkurve is incompatible with AltaiPony, and it slipped me. I did two things to fix this:

1. restricted allowed lightkurve versions (2.x seems incompatible with K2SC light curves, but I'd like to keep that functionality because people still work on those data. The new version also breaks **a lot more** than just `from_mast`, so I think this will not be fixed before major refactoring happens in AltaiPony 2.0.)
2. set up a weekly CI build to notice such failures promptly.
