### 2021-02-25 7497b17
added __version__ and a paragraph on contributions to AltaiPony
