+++ altaipony/version.py
@@ -1,0 +1,3 @@
+# It is important to store the version number in a separate file
+# so that we can read it from setup.py without importing the package
+__version__ = "1.0.1."
