### 2021-02-25 2f12ec4
Updated Github README with instruction for help and contributing
