### 2021-02-25 #2845 @ekaterinailin
As for the other requests:

- `altaipony.__version__` now returns the version
- added a "Contribute to AltaiPony" paragraph on the front page in the docs and in the repo README
