### 2021-02-25 #2845 @ekaterinailin
@afeinstein20 Thanks for the comments!

>  Unfortunately, now I'm running into some issues the way the values are being binned. I've been running everything in this Google Colab notebook. I think if you open it, the error messages should appear as well?

Yes, I can see the error message, and I could indeed reproduce your heatmap error! The pansdas version 1.1.5 is the culprit and I actually fixed that in the requirements back in December (see the [log](https://github.com/ekaterinailin/AltaiPony/commits/master/requirements.txt)), but it was not included in the 1.0.0. version. It should now work with a fresh installation of version 1.0.1 :)

@mrtommyb This should also fix the pip installation problems with lightkurve for the moment.

> Yup this was totally the issue! I thought in your docs you were computing the FFD for TIC 29780677. But I know realize it was for GJ 1243, although I'm still getting fewer points still. This is probably because I've set iterations=50 for the flc.sample_flare_recovery() just for time's sake. Does that sound right to you?

Looking into your Google Colab notebook your results look correct. 16 +/- 2 or so flares is what you'd find in that GJ 1243 light curve. Nothing to do with sample_flare_recovery and the 50 iterations, though. The synthetic flares are stored in `flc.fake_flares`.
