### 2021-03-06 #2845 @mrtommyb
Thanks for your response @ekaterinailin. I think that the tight integration with Lightkurve means that not having the software compatible with Lightkurve 2 means that the software is not entirely functional. I think that it's important to be able to use the current release of Lightkurve, without pinning the installation to an old and unmaintained version of Lightkurve.
