### 2021-03-09 #2845 @dfm
Thanks @mrtommyb!

@ekaterinailin: I agree with Tom that lightkurve compatibility is pretty crucial since it's a major dependency, but also a common element in the workflow of many of your users. I think that it would be worth taking the time to look carefully into what it would take to support lightkurve 2+. Even better, how painful would it be to support both? If it's really infeasible, one option would be to add a prominent version disclaimer to the README and docs, and open an issue describing the steps that would be needed so that folks could contribute.
