### 2021-03-10 #2845 @ekaterinailin
@mrtommyb @dfm Agreed! I have a few deadlines coming up, but I hope that  - if feasible - I can make lightkurve2+ compatibility happen in early April. If that works out I also shall see if I can make the software work for both old and new lightkurve :+1:
