### 2021-04-12 59384ed
trying to fix init with new Lightkurve format
