+++ altaipony/flarelc.py
@@ -11,6 +11,9 @@ from lightkurve import KeplerLightCurve, KeplerTargetPixelFile, TessLightCurve
 from lightkurve.utils import KeplerQualityFlags
 
 from astropy.io import fits
+from astropy.utils.data_info import DataInfo
+from astropy.table import TableColumns, Column
+import astropy.units as u
 
 from .k2scmod import k2sc_lc
 from .altai import (find_flares, find_iterative_median, detrend_savgol)
@@ -72,8 +75,8 @@ class FlareLightCurve(KeplerLightCurve, TessLightCurve):
         Mission identifier, e.g., 'TESS', 'K2' or 'Kepler'.
     cadenceno : array-like
         Cadence number - unique identifier.
-    targetid : int
-        EPIC ID number.
+    object : str
+        target ID.
     ra : float
         RA in deg.
     dec : float
@@ -102,75 +105,81 @@ class FlareLightCurve(KeplerLightCurve, TessLightCurve):
 
 
     """
-    def __init__(self, time=None, flux=None, flux_err=None, time_format=None,
-                 time_scale=None, time_unit=None, centroid_col=None,
-                 centroid_row=None, quality=None, quality_bitmask=None,
-                 channel=None, campaign=None, quarter=None, sector=None, mission=None,
-                 cadenceno=None, targetid=None, ra=None, dec=None, label=None,
-                 meta={}, detrended_flux=None, detrended_flux_err=None,
-                 flux_trends=None, gaps=None, flares=None, flux_unit = None,
-                 primary_header=None, data_header=None, pos_corr1=None,
-                 pos_corr2=None, origin='FLC', fake_flares=None, it_med=None,
-                 pixel_flux=None, pixel_flux_err=None, pipeline_mask=None,
-                 camera=None, ccd=None, saturation=None):
-
+    def __init__(self, time=None, flux=None, flux_err=None, detrended_flux=None,mission=None, detrended_flux_err=None,it_med=None,saturation=None,flares=None, fake_flares=None, meta=None, origin = None, 
+ pixel_flux=None, pixel_flux_err=None, pipeline_mask=None,pos_corr1=None,
+                 pos_corr2=None,**kwargs):
+      
         if mission == 'TESS':
-                TessLightCurve.__init__(self, time=time, flux=flux, flux_err=flux_err,
-                                        time_format=time_format, time_scale=time_scale,
-                                        centroid_col=centroid_col, centroid_row=centroid_row,
-                                        quality=quality, quality_bitmask=quality_bitmask,
-                                        camera=camera, cadenceno=cadenceno, targetid=targetid,
-                                        ra=ra, dec=dec, label=label, meta=meta, sector=sector,
+                TessLightCurve.__init__(self, time=time, flux=flux, flux_err=flux_err, meta=meta, #_required_columns=required_columns,
+ **kwargs
                                         )
-                self.mission = mission
-                self.campaign = None
-                self.quarter = None
+                #self.mission = mission
+                #self.campaign = None
+                #self.quarter = None
         else:
-                KeplerLightCurve.__init__(self, time=time, flux=flux, flux_err=flux_err,
-                                          time_format=time_format, time_scale=time_scale,
-                                          centroid_col=centroid_col, centroid_row=centroid_row,
-                                          quality=quality, quality_bitmask=quality_bitmask,
-                                          channel=channel, campaign=campaign, quarter=quarter,
-                                          mission=mission, cadenceno=cadenceno, targetid=targetid,
-                                          ra=ra, dec=dec, label=label, meta=meta)
-        
[…]
