### 2021-04-12 65fae9b
trying to fix init with new Lightkurve format
