+++ altaipony/flarelc.py
@@ -14,6 +14,8 @@ from astropy.io import fits
 from astropy.utils.data_info import DataInfo
 from astropy.table import TableColumns, Column
 import astropy.units as u
+from astropy.timeseries import TimeSeries
+from astropy.time import Time
 
 from .k2scmod import k2sc_lc
 from .altai import (find_flares, find_iterative_median, detrend_savgol)
@@ -105,62 +107,60 @@ class FlareLightCurve(KeplerLightCurve, TessLightCurve):
 
 
     """
-    def __init__(self, time=None, flux=None, flux_err=None, detrended_flux=None,mission=None, detrended_flux_err=None,it_med=None,saturation=None,flares=None, fake_flares=None, meta=None, origin = None, 
- pixel_flux=None, pixel_flux_err=None, pipeline_mask=None,pos_corr1=None,
-                 pos_corr2=None,**kwargs):
+    #def __init__(self, data, meta=None, flares=None, fake_flares=None,  origin = None, 
+# pixel_flux=None, pixel_flux_err=None, pipeline_mask=None,pos_corr1=None,
+ #                pos_corr2=None, time_format=None, **kwargs):
       
-        if mission == 'TESS':
-                TessLightCurve.__init__(self, time=time, flux=flux, flux_err=flux_err, meta=meta, #_required_columns=required_columns,
- **kwargs
-                                        )
-                #self.mission = mission
-                #self.campaign = None
-                #self.quarter = None
-        else:
-                KeplerLightCurve.__init__(self, time=time, flux=flux, flux_err=flux_err, meta=meta,#_required_columns=required_columns,
- **kwargs)
-
-
-        self.origin = origin
-        for col in [detrended_flux, detrended_flux_err, it_med, saturation]:
-            print(col)
-            if col is None:
-                col = np.full_like(time,np.nan)
+    
+    #    self.origin = origin
+    #    for col in [detrended_flux, detrended_flux_err, it_med, saturation]:
+    #       
+    #        if col is None:
+    #            col = np.full_like(time,np.nan)
        
-        if detrended_flux is None:
-                detrended_flux = np.full_like(time,np.nan)
-        if detrended_flux_err is None:
-                detrended_flux_err = np.full_like(time,np.nan)
-        if it_med is None:
-                it_med = np.full_like(time,np.nan)
-        if saturation is None:
-                saturation = np.full_like(time,np.nan)
-        if pos_corr1 is None:
-                pos_corr1 = np.full_like(time,np.nan)
-        if pos_corr1 is None:
-                pos_corr1 = np.full_like(time,np.nan)
-        # override column checking, otherwise indexing and slicing won't work
-        self._required_columns_enabled = False 
-        self._first_colname = "time"
-      #  for col, name in list(zip([time, flux, flux_err, detrended_flux, detrended_flux_err, it_med, saturation, pos_corr1, pos_corr2], ['time','flux','flux_err','detrended_flux','detrended_flux_err','it_med','saturation', 'pos_corr1','pos_corr2'])):
+      #  if detrended_flux is None:
+      #          detrended_flux = np.full_like(time,np.nan)
+      #  if detrended_flux_err is None:
+      #          detrended_flux_err = np.full_like(time,np.nan)
+      #  if it_med is None:
+      #          it_med = np.full_like(time,np.nan)
+      #  if saturation is None:
+      #          saturation = np.full_like(time,np.nan)
+      #  if pos_corr1 is None:
+      #          pos_corr1 = np.full_like(time,np.nan)
[…]
