### 2021-04-13 83a6fd6
got some flarelc.py tests to work
