+++ altaipony/flarelc.py
@@ -11,11 +11,6 @@ from lightkurve import KeplerLightCurve, KeplerTargetPixelFile, TessLightCurve
 from lightkurve.utils import KeplerQualityFlags
 
 from astropy.io import fits
-from astropy.utils.data_info import DataInfo
-from astropy.table import TableColumns, Column
-import astropy.units as u
-from astropy.timeseries import TimeSeries
-from astropy.time import Time
 
 from .k2scmod import k2sc_lc
 from .altai import (find_flares, find_iterative_median, detrend_savgol)
@@ -29,6 +24,14 @@ from .utils import split_gaps
 
 import time
 LOG = logging.getLogger(__name__)
+
+
+FLARE_COLUMNS = ['istart', 'istop', 'cstart', 'cstop', 'tstart',
+                 'tstop', 'ed_rec', 'ed_rec_err', 'ampl_rec', 
+                  'total_n_valid_data_points', 'dur']
+
+FAKE_FLARE_COLUMNS = ['duration_d', 'amplitude', 'ed_inj', 'peak_time']
+
 
 class FlareLightCurve(KeplerLightCurve, TessLightCurve):
     """
@@ -174,29 +177,74 @@ class FlareLightCurve(KeplerLightCurve, TessLightCurve):
 #        self.add_column(saturation, name="saturation", copy=False, index=len(self._required_columns)+3)
         #self._required_columns = required_columns
         
+        #self._init_flare_table(self, flares=flares, fake_flares=fake_flares)
+        #self._add_tpf_columns(self, pixel_flux=None, pixel_flux_err=None, pipeline_mask=None)
+
+
+    @property
+    def origin(self):
+        try:
+            return self.meta["origin"]
+        except KeyError:
+            self.meta["origin"] = "FLC"
+            return self.meta["origin"]
+
+    @origin.setter
+    def origin(self, origin):
+        self.meta["origin"] = origin 
+    
+    @property
+    def saturation(self):
+        try:
+            return self.meta["saturation"]
+        except KeyError:
+            self.meta["saturation"] = []
+            return self.meta["saturation"]
+
+    @saturation.setter
+    def saturation(self, saturation):
+        self.meta["saturation"] = saturation 
+
+    @property
+    def flares(self) -> pd.DataFrame:
+        try:
+            return self.meta["flares"]
+        except KeyError:
+            self.meta["flares"] = pd.DataFrame(columns=FLARE_COLUMNS)
+            return self.meta["flares"]
+
+    @flares.setter
+    def flares(self, flares):
+        self.meta["flares"] = flares 
+
+
+    @property
+    def fake_flares(self) -> pd.DataFrame:
         
+        try:
+            return self.meta["fake_flares"]
+        except KeyError:
+            self.meta["fake_flares"] = pd.DataFrame(columns=FAKE_FLARE_COLUMNS)
+            return self.meta["fake_flares"]
 
-
+    @fake_flares.setter
+    def fake_flares(self, fake_flares):
+        self.meta["fake_flares"] = fake_flares
 
     def _init_flare_table(self, flares=None, fake_flares=None):
 
-        columns = ['istart', 'istop', 'cstart', 'cstop', 'tstart',
-                   'tstop', 'ed_rec', 'ed_rec_err', 'ampl_rec', 
-                   'total_n_valid_data_points', 'dur']
-
-
         if flares is None:
-            self.flares = pd.DataFrame(columns=columns)
+            self.flares = pd.DataFrame(columns=FLARE_COLUMNS)
         else:
             self.flares = flares
 
         if fake_flares is None:
-            other_columns = ['duration_d', 'amplitude', 'ed_inj', 'peak_time']
[…]
