### 2021-04-13 072ec92
got some flarelc.py tests to work
