+++ altaipony/altai.py
@@ -144,17 +144,17 @@ def find_flares(flc, minsep=3, **kwargs):
     istart = np.array([], dtype='int')
     istop = np.array([], dtype='int')
 
-    if np.isnan(lc.detrended_flux).all():
+    if np.isnan(lc.detrended_flux.value).all():
         raise TypeError('Flare finding only works on de-trended light curves.')
 
     #Now work on periods of continuous observation with no gaps
     for (le,ri) in lc.gaps:
-        error = lc.detrended_flux_err[le:ri]
-        flux = lc.detrended_flux[le:ri]
+        error = lc.detrended_flux_err.value[le:ri]
+        flux = lc.detrended_flux.value[le:ri]
         
-        median = lc.it_med[le:ri]
+        median = lc.it_med.value[le:ri]
         
-        time = lc.time[le:ri]
+        time = lc.time.value[le:ri]
         # run final flare-find on DATA - MODEL
 
         isflare = find_flares_in_cont_obs_period(flux, median, error, **kwargs)
@@ -183,12 +183,12 @@ def find_flares(flc, minsep=3, **kwargs):
     if len(istart) > 0:
         l = [equivalent_duration(lc, i, j, err=True) for (i,j) in zip(istart, istop)]
         ed_rec, ed_rec_err = zip(*l)
-        fl = lc.detrended_flux
-        ampl_rec = [np.max(fl[i:j]) / lc.it_med[i] - 1. for (i,j) in zip(istart,istop)]
-        cstart = lc.cadenceno[istart]
-        cstop = lc.cadenceno[istop]
-        tstart = lc.time[istart]
-        tstop = lc.time[istop]
+        fl = lc.detrended_flux.value
+        ampl_rec = [np.max(fl[i:j]) / lc.it_med.value[i] - 1. for (i,j) in zip(istart,istop)]
+        cstart = lc.cadenceno[istart].value
+        cstop = lc.cadenceno[istop].value
+        tstart = lc.time[istart].value
+        tstop = lc.time[istop].value
 
         lc.flares = lc.flares.append(pd.DataFrame(
                                      {'ed_rec': ed_rec,
@@ -200,7 +200,7 @@ def find_flares(flc, minsep=3, **kwargs):
                                       'cstop': cstop,
                                       'tstart': tstart,
                                       'tstop': tstop,
-                                      'total_n_valid_data_points': lc.flux.shape[0],
+                                      'total_n_valid_data_points': lc.flux.value.shape[0],
                                       'dur': tstop - tstart
                                       }),
                                       ignore_index=True)
@@ -238,10 +238,10 @@ def detrend_savgol(lc, window_length=None, pad=3, printwl=False, **kwargs):
     FlareLightCurve
     '''
     
-    lc = lc[np.where(~np.isnan(lc.flux))]
-    lc.detrended_flux = np.full_like(lc.flux, np.nan)
-    lc.flux_model = np.full_like(lc.flux, np.nan)
-    lc.detrended_flux_err = lc.flux_err
+    lc = lc[np.where(~np.isnan(lc.flux.value))]
+    lc.detrended_flux = np.full_like(lc.flux.value, np.nan)
+    lc["flux_model"] = np.full_like(lc.flux.value, np.nan)
+    lc.detrended_flux_err = lc.flux_err.value
     
     if lc.gaps is None:
         lc = lc.find_gaps()
@@ -255,12 +255,12 @@ def detrend_savgol(lc, window_length=None, pad=3, printwl=False, **kwargs):
     for (wl,le,ri) in gaps:
         
         # Do the iterative sigma clipping
[…]
