+++ altaipony/tests/test_flarelc.py
@@ -61,10 +61,11 @@ def mock_flc(origin='TPF', detrended=False, ampl=1., dur=1):
             'origin' : origin,  'pipeline_mask' : pipeline_mask,
             'pixel_flux' : pixel_flux, 'campaign' : 5, 'ra' : 22.,
             'dec' : 22., 'mission' : 'K2', 'channel' : 55, 
-            'pixel_flux_err' : pixel_flux_err, 'time_format': 'bkjd'}
+            'pixel_flux_err' : pixel_flux_err, 'time_format': 'bkjd',
+            'primary_header':3, 'data_header':2}
 
     if detrended == False:
-        flc = FlareLightCurve(keys)
+        flc = FlareLightCurve(keys, meta=meta)
     else:
         keys["detrended_flux"]=flux
         keys["detrended_flux_err"]=flux_err
@@ -79,7 +80,7 @@ def test_get_saturation():
     # Find saturation overall
     r1 = flc.get_saturation()
     assert r1.saturation[30] == True
-    assert r1.saturation.shape[0] == flc.flux.shape[0]
+    assert r1.saturation.shape[0] == flc.flux.value.shape[0]
     assert (r1.saturation[:30] == False).all()
     assert (r1.saturation[31:] == False).all()
     
@@ -87,7 +88,7 @@ def test_get_saturation():
     # Find saturation with level
     r1 = flc.get_saturation(return_level=True)
     assert r1.saturation[30] == 1e6 / 10093
-    assert r1.saturation.shape[0] == flc.flux.shape[0]
+    assert r1.saturation.shape[0] == flc.flux.value.shape[0]
     assert (r1.saturation[:30] == pytest.approx(500 / 10093, rel=1e-4)) # the LC has some noise added per default, so only approximate results
     assert (r1.saturation[31:] == pytest.approx(500 / 10093, rel=1e-4)) # the LC has some noise added per default, so only approximate results
     
@@ -106,7 +107,7 @@ def test_get_saturation():
     
     
     # .. that is given as a boolean array
-    flc.saturation = np.full(flc.flux.shape[0], True)
+    flc.saturation = np.full(flc.flux.value.shape[0], True)
     r4 = flc.get_saturation()
     assert r4.flares.saturation_f10.iloc[0] == True
     r2 = flc.get_saturation(return_level=True)
@@ -115,7 +116,7 @@ def test_get_saturation():
     assert r3.flares['saturation_f0.01'].iloc[0] == True # throws a warning, too, test that later
     
     # .. that is given as an array of floats
-    flc.saturation = np.full(flc.flux.shape[0], 5.0)
+    flc.saturation = np.full(flc.flux.value.shape[0], 5.0)
     r4 = flc.get_saturation()
     assert r4.flares.saturation_f10.iloc[0] == False
     r2 = flc.get_saturation(return_level=True)
@@ -146,7 +147,7 @@ def test_sample_flare_recovery():
     #test if all injected event are covered in the merged flares:
     assert data.shape[0] == 2
     assert fflc.gaps == [(0, 1000)]
-    assert np.median(fflc.it_med) == pytest.approx(500.005274113832)
+    assert np.median(fflc.it_med.value) == pytest.approx(500.005274113832)
     
     # Custom case
     
@@ -165,8 +166,8 @@ def test_sample_flare_recovery():
     #test if all injected event are covered in the merged flares:
     assert data.shape[0] == 10
     assert fflc.gaps == [(0, 1000)]
-    assert np.median(fflc.it_med) == pytest.approx(500.005274113832/2.)
-    assert flcd.detrended_flux == pytest.approx(flc.flux/2.)
[…]
