+++ altaipony/tests/test_flarelc.py
@@ -270,7 +270,38 @@ def test_find_gaps():
     assert flc.gaps == [(0, 20), (20, 820)]
 
 def test_detrend():
+    # Test K2SC de-trending:
+    flc = mock_flc()
+    try:
+        flc = flc.detrend("k2sc", de_niter=2,)
+        shape = flc.flux.value.shape
+        for att in ["detrended_flux", "detrended_flux_err",
+                    "flux_err", "flux", "time", "quality"]:
+            assert getattr(flc, att).value.shape == shape
+        assert flc.pv[0] == pytest.approx(0.24053657903920067, rel=0.1)
+    except np.linalg.linalg.LinAlgError:
+        warning.warn('Detrending of mock LC failed, this happens.')
+        pass
     
+    # Test K2SC detrending with custom splits    
+    flc = mock_flc()
+    try:
+        flc = flc.detrend("k2sc", de_niter=2, splits=[1.4, 2.2, 5.4])
+        shape = flc.flux.value.shape
+        for att in ["detrended_flux", "detrended_flux_err",
+                    "flux_err", "flux", "time", "quality"]:
+            assert getattr(flc, att).value.shape == shape
+        assert flc.pv[0] == pytest.approx(0.24053657903920067, rel=0.1)
+    except np.linalg.linalg.LinAlgError:
+        warning.warn('Detrending of mock LC failed, this happens.')
+        pass
+
+    #test non TPF derived LC fails
+    flc = mock_flc(origin='LC', detrended=False, ampl=1., dur=1)
+    with pytest.raises(ValueError) as e:
+        flc = flc.detrend("k2sc", de_niter=2,)
+    #test the shapes are the same for all
+    # test that the necessary attributes are kept
     
     # Test SAVGOL detrending
     
