### 2021-04-14 8279a4a
bump lightkurve version
