+++ requirements.txt
@@ -2,7 +2,7 @@ numpy>=1.15.1
 pandas>=1.0, !=1.1.5
 scipy>=1.5.0
 pybind11
-lightkurve>=1.9.1, <=1.11.3
+lightkurve>=2
 astropy>=4.1
 k2sc==1.0.1.4
 emcee
