### 2021-05-10 194f610
added lc to static files, fixed a broken link in docs
