+++ setup.py
@@ -18,7 +18,7 @@ setuptools.setup(
 
     include_package_data=True,
     package_data={
-      'altaipony': ['static/*csv']},
+      'altaipony': ['static/*csv', 'examples/kplr010002792-2010174085026_llc.fits']},
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
