### 2021-05-10 924470a
added lc to static files, fixed a broken link in docs
