### 2021-05-10 0c67a4d
added some docs for light curve detrending, fixed issues #64 and #66
