+++ altaipony/detrend.py
@@ -1,80 +1,364 @@
-'''
-Use this file to keep various detrending methods
-'''
+import time
 
+import copy
 import numpy as np
 import pandas as pd
 
-def MultiBoxcar(time, flux, error, numpass=3, kernel=2.0,
-                sigclip=3, pcentclip=5, returnindx=False,
-                debug=False):
-    '''
-    Boxcar smoothing with multi-pass outlier rejection. Uses both errors
-    and local scatter for rejection Uses Pandas rolling median filter.
+from .altai import find_iterative_median, equivalent_duration
+from .utils import sigma_clip
 
-    Parameters
-    time : 1-d numpy array
-    flux : 1-d numpy array
-    error : 1-d numpy array
-    numpass : int, optional
-        the number of passes to make over the data. (Default = 3)
-    kernel : float, optional
-        the boxcar size in hours. (Default is 2.0)
-        Note: using whole numbers is probably wise here.
-    sigclip : int, optional
-        Number of times the standard deviation to clip points at
-        (Default is 5)
-    pcentclip : int, optional
-        % of data to clip for outliers, i.e. 5= keep 5th-95th percentile
-        (Default is 5)
-    debug : bool, optional
-        used to print out troubleshooting things (default=False)
+from collections import defaultdict
 
-    Returns
+import astropy.units as u
+
+from scipy.interpolate import UnivariateSpline
+from scipy import optimize
+from scipy.fftpack import fft
+
+
+def custom_detrending(lc, spline_coarseness=30, spline_order=3,
+                      savgol1=6., savgol2=3., pad=3):
+    """Custom de-trending for TESS and Kepler 
+    short cadence light curves, including TESS Cycle 3 20s
+    cadence.
+    
+    Parameters:
+    ------------
+    lc : FlareLightCurve
+        light curve that has at least time, flux and flux_err
+    spline_coarseness : float
+        time scale in hours for spline points. 
+        See fit_spline for details.
+    spline_order: int
+        Spline order for the coarse spline fit.
+        Default is cubic spline.
+    savgol1 : float
+        Window size for first Savitzky-Golay filter application.
+        Unit is hours, defaults to 6 hours.
+    savgol2 : float
+        Window size for second Savitzky-Golay filter application.
+        Unit is hours, defaults to 3 hours.
+    pad : 3
+        Outliers in Savitzky-Golay filter are padded with this
+        number of data points. Defaults to 3.
+        
+    Return:
     -------
-    The smoothed light curve model
-    '''
-    flux_i = pd.DataFrame({'flux':flux,'error_i':error,'time_i':time})
-    time_i = np.array(time)
-    error_i = error
-    indx_i = np.arange(len(time)) # for tracking final indx used
-    exptime = np.nanmedian(time_i[1:]-time_i[:-1])
+    FlareLightCurve with detrended_flux attribute
+    """
+# The commented lines will help with debugging, in case the tests in test_detrend.py fail.
 
-    nptsmooth = int(kernel/24.0 / exptime)
+    dt = np.mean(np.diff(lc.time))
+#     plt.figure(figsize=(16,8))
+#     plt.xlim(20,21)
+#     plt.plot(lc.time, lc.flux+2500, c="c", label="original light curve")
+    t0 = time.clock()
+    # fit a spline to the general trends
+    lc1, model = fit_spline(lc, spline_order=spline_order,
+                            spline_coarseness=spline_coarseness)
+    
+    # replace for next step
+    lc1.flux = lc1.detrended_flux
+    t1 = time.clock()
[…]
