### 2021-05-10 76efccf
added some docs for light curve detrending, fixed issues #64 and #66
