+++ docs/source/conf.py
@@ -28,9 +28,9 @@ copyright = '2018, Ekaterina Ilin'
 author = 'Ekaterina Ilin'
 
 # The short X.Y version
-version = ''
+version = '2'
 # The full version, including alpha/beta/rc tags
-release = '0.0.1'
+release = '2.0.0'
 
 
 # -- General configuration ---------------------------------------------------
