### 2021-05-10 83fe980
properly remove MultiBoxCar
