### 2021-05-11 919b4c1
removed detrending with tests, added the detrending function as an experimental feature without test for the moment
