### 2021-05-11 82a68e5
removed detrending with tests, added the detrending function as an experimental feature without test for the moment
