### 2021-05-11 71af0b3
removed a useless test
