### 2021-05-11 037bd76
fixed detrending docs, some tests
