### 2021-05-11 3d385e4
fixed detrending docs, some tests
