### 2021-05-11 9d34b61
added detrending page to docs
