### 2021-05-11 fe95700
updated notebook names in docs
