### 2021-05-12 #2845 @ekaterinailin
Hi @dfm, I finally got back to it. Not quite early April - but now AltaiPony 2.0. is available, test and notebooks seem to run with lightkurve2.0 :)
