### 2021-05-12 #2845 @dfm
@ekaterinailin: Great to hear!

@mrtommyb: Are you be available to revisit your checklist now that we have compatibility with lightkurve 2?
