### 2021-05-19 341c43a
Fixed missing links to Github and email
