+++ README.rst
@@ -96,3 +96,5 @@ Depending on the methods you use, you may also want to cite
 .. _altaipony.readthedocs.io: https://altaipony.readthedocs.io/en/latest/
 .. _notebook: https://github.com/ekaterinailin/AltaiPony/blob/master/notebooks/Getting_Started.ipynb
 .. _docs: https://altaipony.readthedocs.io/en/latest/
+.. _Github: https://github.com/ekaterinailin/AltaiPony/issues/new
+.. _email: eilin@aip.de
