### 2021-05-19 #2845 @ekaterinailin
> Hi @ekaterinailin, I've still working on a the examples, but one very quick thing is that in the Readme you have GitHub and email listed in several places that look like links but they don't link to anywhere.

You're right, thanks! Fixed both links :)
