### 2021-06-09 8ae0d7d
added custom detrending
