+++ altaipony/customdetrend.py
@@ -15,6 +15,8 @@ from scipy.interpolate import UnivariateSpline
 from scipy import optimize
 from scipy.fftpack import fft
 
+import matplotlib.pyplot as plt
+
 
 def custom_detrending(lc, spline_coarseness=30, spline_order=3,
                       savgol1=6., savgol2=3., pad=3):
@@ -48,27 +50,27 @@ def custom_detrending(lc, spline_coarseness=30, spline_order=3,
     """
 # The commented lines will help with debugging, in case the tests in test_detrend.py fail.
 
-    dt = np.mean(np.diff(lc.time))
-#     plt.figure(figsize=(16,8))
-#     plt.xlim(20,21)
-#     plt.plot(lc.time, lc.flux+2500, c="c", label="original light curve")
+    dt = np.mean(np.diff(lc.time.value))
+  #  plt.figure(figsize=(16,8))
+   
+   # plt.plot(lc.time.value, lc.flux.value + 2500, c="c", label="original light curve")
     t0 = time.clock()
     # fit a spline to the general trends
     lc1, model = fit_spline(lc, spline_order=spline_order,
                             spline_coarseness=spline_coarseness)
     
     # replace for next step
-    lc1.flux = lc1.detrended_flux
+    lc1.flux = lc1.detrended_flux.value
     t1 = time.clock()
-#     plt.plot(lc1.time, model+2500, c="r", label="rough trends")
-#     plt.plot(lc1.time, lc1.detrended_flux+500, c="orange", label="rough trends removed")
+    #plt.plot(lc1.time.value, model+2500, c="r", label="rough trends")
+    #plt.plot(lc1.time.value, lc1.detrended_flux.value+500, c="orange", label="rough trends removed")
 
     # removes strong and fast variability on 5 day to 4.8 hours 
     # simple sines are probably because rotational variability is 
     # either weak and transient or strong and persistent on the timescales
     lc2 = remove_sines_iteratively(lc1)
     t2 = time.clock()
-#     plt.plot(lc2time, lc2.detrended_flux-200, label="sines removed")
+    #plt.plot(lc2.time.value, lc2.detrended_flux.value-200, label="sines removed")
     
     # choose a 6 hour window
     w = int((np.rint(savgol1 / 24. / dt) // 2) * 2 + 1)
@@ -82,21 +84,21 @@ def custom_detrending(lc, spline_coarseness=30, spline_order=3,
     # use Savitzy-Golay to iron out the rest
     lc4 = lc3.detrend("savgol", window_length=w, pad=pad)
     t4 = time.clock()
-#     plt.plot(lc4.time, lc4.detrended_flux-800, c="k", label="SavGol applied")
+    #plt.plot(lc4.time.value, lc4.detrended_flux.value-800, c="k", label="SavGol applied")
     
     # find median value
     lc4 = find_iterative_median(lc4)
     t41 = time.clock()
     # replace for next step
-    lc4.flux = lc4.detrended_flux
+    lc4.flux = lc4.detrended_flux.value
     
     # remove exopential fringes that neither spline, 
     # nor sines, nor SavGol 
     # can remove.
     lc5 = remove_exponential_fringes(lc4)
     t5 = time.clock()
-#     plt.plot(lc5.time, lc5.detrended_flux, c="magenta", label="expfunc applied")
-    print(t1-t0, t2-t1, t3-t2, t4-t3, t41-t4, t5-t41, t5-t0)
+    #plt.plot(lc5.time.value, lc5.detrended_flux.value, c="magenta", label="expfunc applied")
[…]
