### 2021-06-09 d941b4c
added custom detrending fixes
