### 2021-06-09 bde4d13
added custom detrending fixes
