### 2021-06-27 #2845 @mrtommyb
Sorry this took me forever to get finished. I'm very happy with the changes made.
