### 2021-06-28 #2845 @ekaterinailin
Thank you @mrtommyb ! I'll look into the warnings - can you recall what numpy version you were working with when you got the warnings?
