### 2021-06-28 #2845 @dfm
@ekaterinailin: I've opened a tiny pull request with an update for the references. Can you take a look at that and once you've merged, please take the following steps:

- Comment `@whedon generate pdf from branch josspaper` on this thread and read through the manuscript to make sure that you're happy with it (it's hard to make changes later!), especially check that your name and affiliation are correct.
- Increment the version number of the software and report that version number back here.
- Create an archived release of that version of the software (using Zenodo or something similar). **Please make sure that the metadata (title and author list) exactly match the paper.** Then report the DOI of the release back to this thread.

Let me know if you have questions or run into any issues!
