### 2021-06-29 fb04e69
Update CHANGELOG.rst
