+++ setup.py
@@ -5,7 +5,7 @@ with open("README.rst", "r") as fh:
 
 setuptools.setup(
     name="altaipony",
-    version="2.0.0",
+    version="2.0.1",
     author="Ekaterina Ilin",
     author_email="eilin@aip.de",
     description="Flare science in Kepler, K2 and TESS light curves",
