+++ README.rst
@@ -1,8 +1,8 @@
 |ci-badge| |docs-badge| |license-badge| |requirements-badge| |zenodo-badge|
 
-.. |zenodo-badge|       image:: https://zenodo.org/badge/DOI/10.5281/zenodo.4095340.svg
-            :target: https://doi.org/10.5281/zenodo.4095340
-
+..  |zenodo-badge| image:: https://zenodo.org/badge/DOI/10.5281/zenodo.5040830.svg
+                    :target: https://doi.org/10.5281/zenodo.5040830
+   
 .. |ci-badge| image:: https://travis-ci.org/ekaterinailin/AltaiPony.svg?branch=master
               :target: https://travis-ci.org/ekaterinailin/AltaiPony
 
