### 2021-06-29 #2845 @ekaterinailin
Thank you for updating and checking the references, @dfm ! I read through the paper - looks complete and correct to me.

The newest version of AltaiPony is now 2.0.1., and the corresponding DOI is [10.5281/zenodo.4095339](https://zenodo.org/record/5040830#.YNsVujqxUn4)
