### 2021-06-29 #2845 @ekaterinailin
Thank you all for getting AltaiPony through the review process, your work and your patience, @dfm @mrtommyb @afeinstein20 :raised_hands: I am very happy with the result :)
