### 2021-06-30 #2845 @ekaterinailin
> @ekaterinailin – your paper is now accepted and published in JOSS

Thanks for wrapping it up so quickly, @arfon !
